Moved log to helpers
